% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_sync_hpc.R
\name{install.packages}
\alias{install.packages}
\title{Install packages locally and remotely}
\usage{
install.packages(package)
}
\arguments{
\item{package}{A vector of package names}
}
\description{
Install packages locally and remotely
}
