% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_sync_hpc.R
\name{install_github}
\alias{install_github}
\title{Install github packages locally and remotely}
\usage{
install_github(repo)
}
\arguments{
\item{repo}{A github repo URL}
}
\description{
Install github packages locally and remotely
}
