% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_sync_hpc.R
\name{install_local}
\alias{install_local}
\title{Synchronize packages between Nuvolos and the HPC Cluster}
\usage{
install_local(path)
}
\arguments{
\item{path}{A local filesystem path (must be given in relative terms, starting with '~/')}
}
\description{
Synchronize packages between Nuvolos and the HPC Cluster
}
