% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_commands.R
\name{run_job_interactive}
\alias{run_job_interactive}
\title{Run job as an interactive R script}
\usage{
run_job_interactive(
  script,
  n_cpus = 4,
  queue = "intq",
  use_mpi = FALSE,
  sync_wait = 60,
  report_freq = 15
)
}
\arguments{
\item{script}{The R script to execute on the cluster. The path to the script should start ~/}

\item{n_cpus}{The number of cpus requested for the operation}

\item{queue}{The queue to use for the job. By default this is intq.}

\item{use_mpi}{Whether to use MPI for running the job (required for multi-node jobs)}

\item{sync_wait}{The amount of time to wait (in seconds) after job completion for the system to synchronize files between Nuvolos and the high performance computing cluster. Larger files warrant larger wait times.}

\item{report_freq}{The frequency of polling job information from the cluster (by default and at least 15 seconds)}
}
\description{
Run job as an interactive R script
}
