% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_commands.R
\name{sbatch}
\alias{sbatch}
\title{Run an R script through SLURM}
\usage{
sbatch(script, n_cpus = 4, queue = "intq", use_mpi = FALSE)
}
\arguments{
\item{script}{The R script to execute on the cluster. The path should start with ~/}

\item{n_cpus}{The number of cpus requested}

\item{mpi}{Whether to use MPI for the job (required for multi-node jobs)}
}
\description{
Run an R script through SLURM
}
