% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_commands.R
\name{slookup_job}
\alias{slookup_job}
\title{Provide detailed information on status of running job}
\usage{
slookup_job(job_id)
}
\arguments{
\item{job_id}{The job to be checked.}
}
\description{
Provide detailed information on status of running job
}
