% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_commands.R
\name{squeue}
\alias{squeue}
\title{Get slurm job status}
\usage{
squeue(user_name = NULL)
}
\arguments{
\item{user_name}{Username at the Nuvolos cluster, leave this empty from a Nuvolos application}
}
\description{
Get slurm job status
}
